+++ docs/console/README.md
@@ -1,0 +1,8 @@
+# Console
+
+The Lula Console is a text-based terminaluser interface that allows users to interact with the OSCAL documents in a more intuitive and visual way.
+
+See the sub-pages for more information on interacting with OSCAL models in the Console.
+
+>[!NOTE]
+>The Console is currently in development and is subject to change.
