+++ docs/console/component-definition.md
@@ -4,9 +4,9 @@ The Component Definition view currently allows for a read-only experience of the
 
 ## Usage
 
-To view an OSCAL Component Definition model in the TUI:
+To view an OSCAL Component Definition model in the Console:
 ```shell
-lula tui -f /path/to/oscal-component.yaml
+lula console -f /path/to/oscal-component.yaml
 ```
 
 ## Keys
@@ -29,4 +29,4 @@ The Component Definition model responds to the following keys for navigation and
 
 ## View
 
-<img align="right" src="../../images/component-defn-tui.gif" alt="component definition tui" style="width:100%; height:auto;">
+<img align="right" src="../../images/component-defn-console.gif" alt="component definition console" style="width:100%; height:auto;">
