--- docs/tui/README.md
@@ -1,8 +1,0 @@
-# TUI
-
-The Terminal User Interface (TUI) is a text-based user interface that allows users to interact with the OSCAL documents in a more intuitive and visual way.
-
-See the sub-pages for more information on interacting with OSCAL models in the TUI.
-
->[!NOTE]
->The TUI is currently in development and is subject to change.
