+++ src/cmd/root.go
@@ -4,11 +4,11 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/defenseunicorns/lula/src/cmd/common"
+	"github.com/defenseunicorns/lula/src/cmd/console"
 	"github.com/defenseunicorns/lula/src/cmd/dev"
 	"github.com/defenseunicorns/lula/src/cmd/evaluate"
 	"github.com/defenseunicorns/lula/src/cmd/generate"
 	"github.com/defenseunicorns/lula/src/cmd/tools"
-	"github.com/defenseunicorns/lula/src/cmd/tui"
 	"github.com/defenseunicorns/lula/src/cmd/validate"
 	"github.com/defenseunicorns/lula/src/cmd/version"
 )
@@ -37,7 +37,7 @@ func init() {
 		validate.ValidateCommand(),
 		evaluate.EvaluateCommand(),
 		generate.GenerateCommand(),
-		tui.TuiCommand(),
+		console.ConsoleCommand(),
 	}
 
 	rootCmd.AddCommand(commands...)
